refactored errors to use thiserror

+++ src/create_tables.rs
@@ -1,16 +1,21 @@
-pub async fn create_tables(pool: &sqlx::PgPool) -> Result<(), sqlx::Error> {
-    sqlx::query!(r#"DROP TABLE IF EXISTS prescribed_drugs;"#)
-        .execute(pool)
-        .await?;
-    sqlx::query!(r#"DROP TABLE IF EXISTS prescriptions;"#)
-        .execute(pool)
-        .await?;
-    sqlx::query!(r#"DROP TYPE IF EXISTS prescriptiontype;"#)
-        .execute(pool)
-        .await?;
+pub async fn create_tables(pool: &sqlx::PgPool, drop: bool) -> Result<(), sqlx::Error> {
+    if drop {
+        sqlx::query!(r#"DROP TABLE IF EXISTS prescribed_drugs;"#)
+            .execute(pool)
+            .await?;
+        sqlx::query!(r#"DROP TABLE IF EXISTS prescriptions;"#)
+            .execute(pool)
+            .await?;
+        sqlx::query!(r#"DROP TYPE IF EXISTS prescriptiontype;"#)
+            .execute(pool)
+            .await?;
+        sqlx::query!(r#"DROP TABLE IF EXISTS doctors;"#)
+            .execute(pool)
+            .await?;
+    }
 
     sqlx::query!(
-        r#"
+    r#"
         DO $$
         BEGIN
             IF NOT EXISTS (SELECT 1 FROM pg_type WHERE typname = 'prescriptiontype') THEN
